requirements.txt
openai
langchain
python-dotenv
requests
azure-identity
azure-ai-textanalytics
